refactor(core): Extract multi-root display path helpers into rootDisplayPath module
intent(packager): keep packager.ts lean by moving the multi-root label/display-path helpers into their own module
decision(rootDisplayPath): expose only buildRootLabels and joinDisplayPath; keep toDisplayPath and the label-dedup helpers private
decision(tests): add direct unit tests for the extracted helpers, which were previously only exercised via the multiRootSpec integration test
