+++ src/core/packager.ts
@@ -15,6 +15,7 @@ import { calculateMetrics, createMetricsTaskRunner } from './metrics/calculateMe
 import { loadTokenCountCache, saveTokenCountCache } from './metrics/tokenCountCache.js';
 import { prefetchSortData, sortOutputFiles } from './output/outputSort.js';
 import { produceOutput } from './packager/produceOutput.js';
+import { buildRootLabels, joinDisplayPath } from './packager/rootDisplayPath.js';
 import type { SuspiciousFileResult } from './security/securityCheck.js';
 import { validateFileSafety } from './security/validateFileSafety.js';
 import type { PackSkillParams } from './skill/packSkill.js';
@@ -64,100 +65,6 @@ export interface PackOptions {
   skillProjectName?: string;
   skillSourceUrl?: string;
 }
-
-const toDisplayPath = (filePath: string): string => filePath.replaceAll(path.win32.sep, path.posix.sep);
-
-const isCwdRelativePath = (relativePath: string): boolean =>
-  relativePath !== '' &&
-  relativePath !== '..' &&
-  !relativePath.startsWith(`..${path.sep}`) &&
-  !path.isAbsolute(relativePath);
-
-const getDuplicateLabels = (labels: string[]): Set<string> => {
-  const counts = new Map<string, number>();
-  for (const label of labels) {
-    counts.set(label, (counts.get(label) ?? 0) + 1);
-  }
-
-  const duplicates = new Set<string>();
-  for (const [label, count] of counts) {
-    if (count > 1) {
-      duplicates.add(label);
-    }
-  }
-  return duplicates;
-};
-
-const uniquifyLabelsWithSuffixes = (labels: string[]): string[] => {
-  const seen = new Set<string>();
-  return labels.map((label, index) => {
-    const baseLabel = label || `root-${index + 1}`;
-    let candidate = baseLabel;
-    let suffix = 2;
-    while (seen.has(candidate)) {
-      candidate = `${baseLabel}-${suffix}`;
-      suffix++;
-    }
-    seen.add(candidate);
-    return candidate;
-  });
-};
-
-const buildRootLabels = (rootDirs: string[], cwd: string): string[] => {
-  const resolvedCwd = path.resolve(cwd);
-  const resolvedRootDirs = rootDirs.map((rootDir) => path.resolve(rootDir));
-  const labelCandidates = resolvedRootDirs.map((rootDir) => {
-    const relativeRootDir = path.relative(resolvedCwd, rootDir);
-    if (isCwdRelativePath(relativeRootDir)) {
-      const label = toDisplayPath(relativeRootDir);
-      return {
-        label,
-        segments: label.split('/').filter(Boolean),
-      };
-    }
-
-    return {
-      label: toDisplayPath(path.basename(rootDir) || rootDir),
-      segments: undefined,
-    };
-  });
-  const labels = labelCandidates.map(({ label }) => label);
-
-  if (new Set(labels).size === labels.length) {
-    return labels;
-  }
-
-  const duplicateLabels = getDuplicateLabels(labels);
-  const maxDepth = Math.max(...labelCandidates.map(({ segments }) => segments?.length ?? 1), 1);
-  let candidates = labels;
-
-  for (let depth = 1; depth <= maxDepth; depth++) {
-    candidates = labels.map((label, index) => {
-      if (!duplicateLabels.has(label)) {
-        return label;
-      }
-
-      const segments = labelCandidates[index]?.segments;
-      if (!segments) {
-        return label || `root-${index + 1}`;
-      }
-
-      return segments.slice(-depth).join('/') || label || `root-${index + 1}`;
-    });
-
-    if (new Set(candidates).size === candidates.length) {
-      return candidates;
-    }
-  }
-
-  return uniquifyLabelsWithSuffixes(candidates);
-};
-
-const joinDisplayPath = (rootLabel: string, filePath: string): string => {
-  const normalizedRootLabel = toDisplayPath(rootLabel).replace(/^\/+|\/+$/g, '') || 'root';
-  const normalizedFilePath = toDisplayPath(filePath).replace(/^\/+/, '');
-  return normalizedFilePath ? `${normalizedRootLabel}/${normalizedFilePath}` : normalizedRootLabel;
-};
 
 export const pack = async (
   rootDirs: string[],
